+++ test/autocomplete_test.rb
@@ -56,6 +56,16 @@ class AutocompleteTest < Minitest::Test
     assert_search "dark grey", ["Dark Grey"], fields: [{name: :word_start}]
   end
 
+  def test_word_start_exact
+    store_names ["Back Scratcher", "Backpack"]
+    assert_order "back", ["Back Scratcher", "Backpack"], fields: [{name: :word_start}]
+  end
+
+  def test_word_start_exact_martin
+    store_names ["Martina", "Martin"]
+    assert_order "martin", ["Martin", "Martina"], fields: [{name: :word_start}]
+  end
+
   # TODO find a better place
 
   def test_exact